requirements.txt
schedule
psutil
pillow
